Create Remote File Download
It looks like you've added a comment to the code snippet, which is a good practice to document your code and make it easier for other developers to understand. In this case, the comment explains what each section of the code is doing:

The first section checks if the "download" button was clicked and gets the remote file URL from the input field.
The second section defines the local file path and name where the downloaded file will be saved.
The third section downloads the remote file using file_get_contents() and saves it locally using file_put_contents().
The final section displays a message to the user that the file was downloaded successfully.
Overall, the code is well-organized and easy to understand, which makes it a good example for other developers who need to perform a similar task.